Feat: division 데이터 형번환 추가

+++ Calculator.playground/Contents.swift
@@ -3,7 +3,7 @@ protocol AbstractOperation: AnyObject {
     func add(_ num1: Int, _ num2: Int) -> Int
     func minus(_ num1: Int, _ num2: Int) -> Int
     func multiply(_ num1: Int, _ num2: Int) -> Int
-    func division(_ num1: Double, _ num2: Double) -> Double?
+    func division(_ num1: Int, _ num2: Int) -> Double?
     func remainder(_ num1: Int, _ num2: Int) -> Int?
 }
 
@@ -18,7 +18,7 @@ class Calculator: AbstractOperation {
     let divisionResult = DivideOperation()
     let remainderResult = RemainderOperation()
     
-    
+
     
     func add(_ num1: Int, _ num2: Int) -> Int {
         return addResult.add(num1, num2)
@@ -35,7 +35,7 @@ class Calculator: AbstractOperation {
     
     // 분모가 0인 경우 대비
     
-    func division(_ num1: Double, _ num2: Double) -> Double? {
+    func division(_ num1: Int, _ num2: Int) -> Double? {
         guard num2 != 0 else { return nil }
             return divisionResult.division(num1, num2)
     }
@@ -71,8 +71,8 @@ class MultiplyOperation {
 }
     
 class DivideOperation {
-    func division(_ num1: Double, _ num2: Double) -> Double {
-        return num1 / num2
+    func division(_ num1: Int, _ num2: Int) -> Double {
+        return Double(num1) / Double(num2)
     }
 }
